Refactored algebra
