--- base/shared/src/main/scala/scalaz/algebra/MonoidClass.scala
@@ -1,6 +1,0 @@
-package scalaz
-package algebra
-
-trait MonoidClass[A] extends SemigroupClass[A] {
-  def empty: A
-}
