--- base/shared/src/main/scala/scalaz/algebra/SemigroupClass.scala
@@ -1,6 +1,0 @@
-package scalaz
-package algebra
-
-trait SemigroupClass[A] {
-  def append(a1: A, a2: => A): A
-}
