--- base/shared/src/main/scala/scalaz/algebra/SemigroupInstances.scala
@@ -1,4 +1,0 @@
-package scalaz
-package algebra
-
-trait SemigroupInstances {}
