+++ base/shared/src/main/scala/scalaz/algebra/monoid.scala
@@ -1,6 +1,10 @@
 package scalaz
 package algebra
 
+trait MonoidClass[A] extends SemigroupClass[A] {
+  def empty: A
+}
+
 trait MonoidInstances {
   implicit val stringMonoid: Monoid[String] = instanceOf(new MonoidClass[String] {
     def append(a1: String, a2: => String) = a1 + a2
