+++ base/shared/src/main/scala/scalaz/algebra/ord.scala
@@ -1,6 +1,8 @@
 package scalaz
 package algebra
 
+import scala.language.experimental.macros
+
 import core.EqClass
 
 sealed abstract class Ordering extends Product with Serializable
@@ -16,4 +18,10 @@ trait OrdClass[A] extends EqClass[A] {
   def >(a: A, b: A): Boolean  = comp(a, b) eq GT
   def >=(a: A, b: A): Boolean = comp(a, b) ne LT
   def equal(a: A, b: A)       = comp(a, b) eq EQ
+}
+
+trait OrdSyntax {
+  implicit final class ToOrdOps[A](a: A) {
+    def comp(f: A)(implicit ev: Ord[A]): Ordering = macro meta.Ops.ia_1
+  }
 }
